+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Removed
+- `XRPTransaction.account` and `XRPTransaction.sourceTag` were deprecated for two releases and have been removed.
+   Please use the X-address encoded field `sourceXAddress` instead.
+- `XRPPayment.destination` and `XRPPayment.destinationTag` were deprecated for two releases and have been removed.
+   Please use the X-address encoded field `destinationXAddress` instead.
+
 ## 3.2.2 - 2020-06-18
 
 ### Fixed